+++ Day 2/script.js
@@ -1,32 +1,13 @@
-const student = {
-  name: "Anas",
-  marks: {
-    math: 85,
-    physics: 78,
-    chemistry: 92,
-    english: 74
+const marks = [75, 89, 92, 67, 81];
+
+let highest = marks[0]; // assume first element is highest
+
+for (let i = 1; i < marks.length; i++) {
+  if (marks[i] > highest) {
+    highest = marks[i];
   }
-};
+}
 
-// Convert marks object to array
-let marksArray = Object.values(student.marks);
-
-// Total marks
-let total = marksArray.reduce((acc, val) => acc + val, 0);
-
-// Average marks
-let average = total / marksArray.length;
-
-// Highest marks
-let highest = Math.max(...marksArray);
-
-console.log("Student:", student.name);
-console.log("Total Marks:", total);
-console.log("Average Marks:", average);
 console.log("Highest Marks:", highest);
-// then write "node script.js" in terminal and press enter
-// Output:
-// Student: Aman
-// Total Marks: 329
-// Average Marks: 82.25
-// highest Marks: 92
+// for running the code write "node script.js" and press enter in the terminal
+// Output: Highest Marks: 92
